+++ backend/ElectroManage.Domain/Entites/Sucursal/Register.cs
@@ -4,8 +4,8 @@ namespace ElectroManage.Domain.Entites.Sucursal;
 public class Register : Entity<long>
 {
     public DateTime Date { get; set; }
-    public decimal Cost { get; set; }
-    public decimal Consumption { get; set; }
+    public double Cost { get; set; }
+    public double Consumption { get; set; }
     public long CompanyId { get; set; }
     public Company Company { get; set; } = null!;
 }
